+++ src/main.cpp
@@ -6,7 +6,7 @@
 #include "handlers/admin_v1_configs.hpp"
 #include "handlers/admin_v1_configs_delete.hpp"
 #include "handlers/configs_values.hpp"
-#include "handlers/delete_variable/view.hpp"
+#include "handlers/delete-variables_uuid/view.hpp"
 #include "userver/clients/dns/component.hpp"
 #include "userver/clients/http/component.hpp"
 #include "userver/testsuite/testsuite_support.hpp"
@@ -27,7 +27,7 @@ int main(int argc, char *argv[]) {
           .Append<service_handlers::configs_values::post::Handler>()
           .Append<service_handlers::admin_v1_configs::post::Handler>()
           .Append<service_handlers::admin_v1_configs_delete::post::Handler>()
-          .Append<service_handlers::delete_variable::Handler>()
+          .Append<service_handlers::variables_uuid::del::Handler>()
           .Append<userver::components::HttpClient>()
           .Append<userver::server::handlers::TestsControl>();
   return userver::utils::DaemonMain(argc, argv, component_list);
